Cheange BMI2

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weightInKg, double lengthInMetres1) {
+    public int calculate(int weightInKg, double lengthInMetres1) {
         double index;
         index = weightInKg / (lengthInMetres1 * 2);
         return (int) index;
